Create zwave-rgbw-bulb.groovy
Creation of a new "zwave-rgbw-bulb" to be the eventual replacement of the "rgbw-light" DTH which is not being maintained with recent changes and is for the most part a duplicate of "aeon-led-bulb".  Naming is also more consistant with other DTH's (zll-rgbw-bulb, zigbee-rgbw-bulb, etc).   Re: https://github.com/SmartThingsCommunity/SmartThingsPublic/pull/3599#issuecomment-432804367

Source code is based on the aeon as it was most up to date.  Modifications made:
- Updated copyright and author
- Merged all fingerprints from both "aeon-led-bulb" and "rgbw-light" into this DTH
- Added metadata for ocfDeviceType, mnmn, and vid with proper identifiers
- Switched standard tile over to a multiAttributeTile to more closely match "zll-rgbw-bulb"
- Removed tiles no longer needed after conversion to mutliAttributeTile
- Added capability "Light"

Needs full testing but since code is 97% from existing working DTH it should be minimal.  Also would be nice once approved to have this added to locally run devices like other "generic" Z-Wave and ZigBee DTH's.  Eventually I would think the aeon-rgbw-bulb and rgbw-light can both be replaced by this.

NOTE: Namespace needs to be changed before submission to SmartThings public GitHub